requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
ml-collections>=0.1.1
absl-py>=1.4.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
monai>=1.2.0
nibabel>=5.0.0

# Augmentation & Preprocessing
albumentations>=1.3.0
opencv-python>=4.7.0

# Metrics & Evaluation
torchmetrics>=0.11.0
scikit-image>=0.20.0

# Logging & Experiment Tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
Pillow>=9.5.0

# ODE Solvers
torchdiffeq>=0.2.3

# Utils
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0

# Optional: For better performance
# ninja>=1.11.0
# flash-attn>=2.0.0
